Диплом 2 часть

+++ ewm-service/src/main/java/ru/practicum/explorewithme/server/adminAPI/servise/user/AdminUserService.java
@@ -6,7 +6,7 @@ import ru.practicum.explorewithme.server.dto.user.UserDto;
 import java.util.List;
 
 public interface AdminUserService {
-    List<UserDto> getAll(List<Integer> ids, Integer from, Integer size);
+    List<UserDto> getAll(List<Long> ids, Integer from, Integer size);
 
     UserDto save(NewUserRequest newUserRequest);
